svg/types/scripted/SVGGeometryElement.getTotalLength-02.svg

<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:xlink="http://www.w3.org/1999/xlink"
     xmlns:html="http://www.w3.org/1999/xhtml">
  <title>When SVGGeometryElement.getTotalLength is called with an element that is not in the document, throw exception</title>
  <html:link rel="help" href="https://svgwg.org/svg2-draft/types.html#__svg__SVGGeometryElement__getTotalLength"/>
  <html:script src="/resources/testharness.js"/>
  <html:script src="/resources/testharnessreport.js"/>
  <script><![CDATA[
  test(function() {
    var pathElement = document.createElementNS("http://www.w3.org/2000/svg", "path");
    pathElement.setAttribute("d", 'M0,20 L400,20 L640,20');
    assert_approx_equals(pathElement.getTotalLength(), 640, 1e-5);
  }, document.title + " with SVGPathElement");

  test(function() {
    var pathElement = document.createElementNS("http://www.w3.org/2000/svg", "path");
    assert_throws_dom("InvalidStateError", function() { pathElement.getTotalLength() });
  }, document.title + " with SVGPathElement (no 'd' attribute)");

  test(function() {
    var rectElement = document.createElementNS("http://www.w3.org/2000/svg", "rect");
    rectElement.setAttribute("rx", 0);
    rectElement.setAttribute("ry", 0);
    rectElement.setAttribute("width", 200);
    rectElement.setAttribute("height", 300);
    assert_approx_equals(rectElement.getTotalLength(), 1000, 1e-5);
  }, document.title + " with SVGRectElement");

  test(function() {
    var circleElement = document.createElementNS("http://www.w3.org/2000/svg", "circle");
    circleElement.setAttribute("r", 10);
    assert_approx_equals(circleElement.getTotalLength(), 62.8, 1e-1);
  }, document.title + " with SVGCircleElement");
  ]]></script>
</svg>
